refactor(game-client): remove obsolete components from Explore Alphabet page
- Deleted unused components: AlphabetLessonsList, ContextAwareGreeting, and CardButton, streamlining the Explore Alphabet page.
- Updated the page structure to utilize the ExploreAlphabetPage component from the UI library, enhancing maintainability and clarity.
- Adjusted global styles for improved consistency across the application.
