+++ apps/game-client/src/ui/experiences/explore/components/content-card/content-card-section.tsx
@@ -1,0 +1,18 @@
+import { cn } from '@kartuli/ui/utils/cn';
+
+export function ContentCardSection({ children }: Readonly<{ children: React.ReactNode }>) {
+  return (
+    <div
+      className={cn(
+        //
+        'w-full',
+        'px-4',
+        'py-2',
+        'gap-4',
+        'flex',
+      )}
+    >
+      {children}
+    </div>
+  );
+}
